requirements.txt
requests>=2.28.0
requests-toolbelt>=1.0.0
python-dotenv>=1.0.0

# Dev / testing
coverage[toml]>=7.0.0
